requirements.txt
asgiref
boto3
botocore
cffi
click
cryptography
fastapi
fastapi-pagination
h11
pycparser
pydantic
PyJWT
starlette
typing-extensions
uvicorn
pytest
pyqtwebengine
httpx
SQLAlchemy
python-dotenv
python-multipart
psycopg2-binary
email-validator
pydantic-settings
psutil
